Separated commands and completers
